+++ src/apis/auth/types/index.ts
@@ -1,0 +1,7 @@
+/**
+ * @descript 회원가입할 때 폼 상태 타입
+ */
+export interface ISignUpRequest {
+  email: string;
+  password: string;
+}
